Value handler works for array types but not regular
Attaching deserializer to the JavaType object passed into the ObjectMapper because the KotlinModule only
allows deserializers to be registered for basic Java types and I want to register
deserializers for the same parameterized type parameterized with different sets of values. When I do this
and deserialize an array of my parameterized types it works: my values are deserialized correctly. But
when I do the same thing and try to deserialize values that are not in an array the ObjectMapper doesn't
find my deserializer and deserialization fails.